# Backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Project Rebound Game Control Plane API
  version: 1.3.0
  description: Public control-plane API. Internal Relay APIs are marked with x-internal.
servers:
  - url: https://api.example.com
paths:
...
  /connectServer:
    post:
      operationId: connectMetaServerCompatibility
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      description: Compatibility endpoint used only through MetaTunnel. Any playerId or loginToken field is ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MetaSessionRequest"}
      responses:
        "200":
          description: Compatibility Gate response. MetaTunnel rewrites endpoint to its loopback listener.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaConnectServerResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "429": {$ref: "#/components/responses/RateLimited"}
...
  /v1/meta/notifications:
    get:
      operationId: listMetaNotifications
      tags: [MetaServer]
      parameters:
        - in: query
          name: locale
          schema: {type: string, maxLength: 16}
      responses:
        "200":
          description: Active notifications in the requested locale.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaNotificationListResponse"}
  /v1/meta/sessions:
    post:
      operationId: createMetaSession
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MetaSessionRequest"}
      responses:
        "201":
          description: A single-use, short-lived Gate ticket.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaSessionResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /v1/users/me/meta-profile:
    get:
      operationId: getMyMetaProfile
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      responses:
        "200":
          description: Authenticated player's Meta profile.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaProfileResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /v1/users/me/loadouts:
    get:
      operationId: listMyMetaLoadouts
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      responses:
        "200":
          description: Authenticated player's role loadouts.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaLoadoutListResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /v1/users/me/loadouts/{role_id}:
    parameters:
      - {$ref: "#/components/parameters/MetaRoleID"}
    get:
      operationId: getMyMetaLoadout
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      responses:
        "200":
          description: Role loadout and current revision.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaLoadoutResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
    put:
      operationId: updateMyMetaLoadout
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MetaLoadoutUpdateRequest"}
      responses:
        "200":
          description: Updated role loadout with incremented revision.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaLoadoutResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
  /v1/meta/parties:
    post:
      operationId: createMetaParty
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MetaPartyCreateRequest"}
      responses:
        "201":
          description: Created party.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaPartyResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
  /v1/meta/parties/{party_id}:
    parameters:
      - {$ref: "#/components/parameters/MetaPartyID"}
    get:
      operationId: getMetaParty
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      description: Returns the party only when the authenticated player is a member.
      responses:
        "200":
          description: Party state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaPartyResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /v1/meta/parties/{party_id}/presence:
    parameters:
      - {$ref: "#/components/parameters/MetaPartyID"}
    post:
      operationId: setMetaPartyPresence
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MetaPartyPresenceRequest"}
      responses:
        "200":
          description: Updated party.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaPartyResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /v1/meta/matchmaking/tickets:
    post:
      operationId: createMetaMatchTicket
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MetaMatchTicketCreateRequest"}
      responses:
        "202":
          description: Queued match ticket. No P2P fallback is implied.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaMatchTicketResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
  /v1/meta/matchmaking/tickets/{ticket_id}:
    parameters:
      - {$ref: "#/components/parameters/MetaTicketID"}
    get:
      operationId: getMetaMatchTicket
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      description: Ticket ownership is derived from the access token; cross-player access is hidden as not found.
      responses:
        "200":
          description: Current ticket state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaMatchTicketResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
    delete:
      operationId: cancelMetaMatchTicket
      tags: [MetaServer]
      security:
        - playerAccessToken: []
      responses:
        "204": {description: Ticket cancelled.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /internal/v1/meta/matches/{match_id}/players/{player_id}/loadout:
    get:
      operationId: getAssignedMetaPlayerLoadout
      tags: [MetaServer Internal]
      x-internal: true
      security:
        - gameServerToken: []
      parameters:
        - {$ref: "#/components/parameters/MetaMatchID"}
        - {$ref: "#/components/parameters/PlayerID"}
        - {$ref: "#/components/parameters/GameServerIdentity"}
      description: Requires meta.loadouts.read scope and an active match assigned to this Game Server.
      responses:
        "200":
          description: Match-bound loadout snapshot.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaMatchPlayerLoadoutResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /internal/v1/meta/matches/{match_id}/players/{player_id}/connected:
    post:
      operationId: markAssignedMetaPlayerConnected
      tags: [MetaServer Internal]
      x-internal: true
      description: Requires meta.matches.connect scope and the assigned active match/player.
      security:
        - gameServerToken: []
      parameters:
        - {$ref: "#/components/parameters/MetaMatchID"}
        - {$ref: "#/components/parameters/PlayerID"}
        - {$ref: "#/components/parameters/GameServerIdentity"}
      responses:
        "204": {description: Player connection recorded.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /internal/v1/meta/matches/{match_id}/completed:
    post:
      operationId: completeAssignedMetaMatch
      tags: [MetaServer Internal]
      x-internal: true
      description: Requires meta.matches.complete scope and an active match assigned to this Game Server.
      security:
        - gameServerToken: []
      parameters:
        - {$ref: "#/components/parameters/MetaMatchID"}
        - {$ref: "#/components/parameters/GameServerIdentity"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MetaMatchCompleteRequest"}
      responses:
        "204": {description: Match completed and Game Server returned to READY.}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /internal/v1/meta/battlelog/reports/{report_id}:
    put:
      operationId: submitBattleLogReport
      tags: [MetaServer Internal]
      x-internal: true
      description: |
        Requires meta.battlelog.write scope. The report is idempotent per
        Game Server and report_id. A matching active assignment is authoritative;
        otherwise the report is retained as a non-official standalone record.
      security:
        - gameServerToken: []
      parameters:
        - {$ref: "#/components/parameters/BattleLogReportID"}
        - {$ref: "#/components/parameters/GameServerIdentity"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BattleLogSubmitRequest"}
      responses:
        "200":
          description: Idempotent replay of an existing report with the same SHA-256.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BattleLogSubmitResponse"}
        "201":
          description: BattleLog report stored and normalized.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BattleLogSubmitResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403":
          description: Scope, Game Server, or assigned-match check failed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422":
          description: Snapshot schema or required server-side evidence is invalid.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /v1/admin/meta/overview:
    get:
      operationId: adminGetMetaOverview
      tags: [Admin MetaServer]
      security:
        - adminAccessToken: []
      responses:
        "200":
          description: MetaServer operational counters.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminMetaOverviewResponse"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /v1/admin/meta/players/{player_id}/loadouts:
    get:
      operationId: adminListMetaPlayerLoadouts
      tags: [Admin MetaServer]
      security:
        - adminAccessToken: []
      parameters:
        - {$ref: "#/components/parameters/PlayerID"}
      responses:
        "200":
          description: Player loadouts for authorized support review.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaLoadoutListResponse"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /v1/admin/meta/players/{player_id}/loadouts/{role_id}:
    put:
      operationId: adminUpdateMetaPlayerLoadout
      tags: [Admin MetaServer]
      security:
        - adminAccessToken: []
          adminStepUpToken: []
      parameters:
        - {$ref: "#/components/parameters/PlayerID"}
        - {$ref: "#/components/parameters/MetaRoleID"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminMetaLoadoutUpdateRequest"}
      responses:
        "200":
          description: Updated and audited loadout.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaLoadoutResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
  /v1/admin/meta/matches:
    get:
      operationId: adminListMetaMatches
      tags: [Admin MetaServer]
      security:
        - adminAccessToken: []
      responses:
        "200":
          description: Most recently updated matches.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminMetaMatchListResponse"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /v1/admin/meta/matches/{match_id}/cancel:
    post:
      operationId: adminCancelMetaMatch
      tags: [Admin MetaServer]
      security:
        - adminAccessToken: []
          adminStepUpToken: []
      parameters:
        - {$ref: "#/components/parameters/MetaMatchID"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminReasonRequest"}
      responses:
        "204": {description: "Match cancelled, reservation released, and action audited."}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /v1/admin/meta/playlists/{slug}:
    put:
      operationId: adminUpsertMetaPlaylist
      tags: [Admin MetaServer]
      security:
        - adminAccessToken: []
          adminStepUpToken: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: {type: string, pattern: "^[a-z0-9][a-z0-9-]{0,63}$"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminMetaPlaylistUpsertRequest"}
      responses:
        "200":
          description: Playlist upserted and audited.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaPlaylistResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /v1/admin/meta/notifications/{notification_id}:
    put:
      operationId: adminUpsertMetaNotification
      tags: [Admin MetaServer]
      security:
        - adminAccessToken: []
          adminStepUpToken: []
      parameters:
        - in: path
          name: notification_id
          required: true
          schema: {type: string, maxLength: 64}
          description: Use new to allocate a new notification ID.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminMetaNotificationUpsertRequest"}
      responses:
        "200":
          description: Notification upserted and audited.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MetaNotificationResponse"}
        "400": {$ref: "#/components/responses/InvalidRequest"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
...
components:
  securitySchemes:
    playerAccessToken:
      type: http
      scheme: bearer
      bearerFormat: EdDSA JWT
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: Static Admin Token
      description: Machine-only token for internal operational routes. It is not accepted by the Admin Web API.
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: EdDSA Administrator JWT
      description: Short-lived human administrator token issued only after Turnstile, password, and MFA validation.
    adminStepUpToken:
      type: apiKey
      in: header
      name: X-Admin-Step-Up
      description: Short-lived MFA proof bound to the current administrator and session.
    gameServerRegistrationToken:
      type: http
      scheme: bearer
      bearerFormat: One-Time Instance-Bound Registration Token
      description: Single-use token created by an authorized administrator and bound to the request instance_id.
    gameServerToken:
      type: http
      scheme: bearer
      bearerFormat: Opaque Game Server Token
      description: Short-lived node token. Signed endpoints additionally require the Game Server signature headers.
    gameServerSignature:
      type: apiKey
      in: header
      name: X-Game-Server-Signature
      description: Base64url Ed25519 signature over the PR-GAME-SERVER-V1 canonical request. Timestamp, nonce, certificate fingerprint, credential generation, node ID, body hash, and token hash are bound into the signature.
    relayBootstrapToken:
      type: http
      scheme: bearer
      bearerFormat: One-Time Relay Bootstrap Token
    relayNodeToken:
      type: http
...
  parameters:
...
    PlayerID:
      in: path
      name: player_id
      required: true
      schema:
        type: string
...
    MetaRoleID:
      in: path
      name: role_id
      required: true
      schema:
        type: string
        pattern: "^[A-Za-z0-9_.:-]{1,128}$"
    MetaPartyID:
      in: path
      name: party_id
      required: true
      schema:
        type: string
        pattern: "^mp_[A-Za-z0-9_-]+$"
    MetaTicketID:
      in: path
      name: ticket_id
      required: true
      schema:
        type: string
        pattern: "^mt_[A-Za-z0-9_-]+$"
    MetaMatchID:
      in: path
      name: match_id
      required: true
      schema:
        type: string
        pattern: "^mm_[A-Za-z0-9_-]+$"
    BattleLogReportID:
      in: path
      name: report_id
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
    GameServerIdentity:
      in: header
      name: X-Game-Server-Id
      required: true
      schema:
        type: string
      description: Must identify the same Game Server encoded in the scoped bearer token.
  responses:
    InvalidRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Invalid, expired, reused, or revoked session token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: An authentication rate limit was exceeded. Retry-After is returned when applicable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found or not exposed on this network.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The requested operation conflicts with the current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    MetaNotificationListResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              type: array
              items: {$ref: "#/components/schemas/MetaNotification"}
        request_id: {type: string}
    MetaSessionRequest:
      type: object
      additionalProperties: false
      properties:
        client_version: {type: string, minLength: 1, maxLength: 64}
        version:
          type: string
          minLength: 1
          maxLength: 64
          description: Compatibility alias for client_version.
        protocol_version: {type: integer, minimum: 1}
        platform: {type: string, maxLength: 64}
        playerId:
          type: string
          deprecated: true
          description: Ignored. Identity is always derived from the access token.
        loginToken:
          type: string
          deprecated: true
          description: Ignored. Authorization is accepted only through the bearer header.
...
    MetaSessionResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/MetaSessionData"}
        request_id: {type: string}
    MetaConnectServerResponse:
      type: object
      additionalProperties: false
      required: [error, userId, aceId, gateToken, endpoint]
      properties:
        error: {type: integer, const: 0}
        userId: {type: string}
        aceId: {type: string}
        gateToken: {type: string, writeOnly: true}
        endpoint: {type: string}
...
    MetaProfileResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/MetaProfile"}
        request_id: {type: string}
...
    MetaLoadoutUpdateRequest:
      type: object
      additionalProperties: false
      required: [snapshot, revision]
      properties:
        snapshot: {type: object, additionalProperties: true}
        revision:
          type: integer
          format: int64
          minimum: 0
          description: Current revision, or zero when creating a role loadout.
...
    MetaLoadoutResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/MetaLoadout"}
        request_id: {type: string}
    MetaLoadoutListResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              type: array
              items: {$ref: "#/components/schemas/MetaLoadout"}
        request_id: {type: string}
    MetaPartyCreateRequest:
      type: object
      additionalProperties: false
      required: [mode, region, client_version]
      properties:
        mode: {type: string, minLength: 1, maxLength: 64}
        region: {type: string, minLength: 1, maxLength: 64}
        client_version: {type: string, minLength: 1, maxLength: 64}
...
    MetaPartyResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/MetaParty"}
        request_id: {type: string}
...
    MetaPartyPresenceRequest:
      type: object
      additionalProperties: false
      required: [presence]
      properties:
        presence: {type: string, enum: [ONLINE, AWAY, IN_GAME, OFFLINE]}
    MetaMatchTicketCreateRequest:
      type: object
      additionalProperties: false
      required: [mode, region, client_version]
      properties:
        party_id: {type: string}
        mode: {type: string, minLength: 1, maxLength: 64}
        region: {type: string, minLength: 1, maxLength: 64}
        client_version: {type: string, minLength: 1, maxLength: 64}
...
    MetaMatchTicketResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/MetaMatchTicket"}
        request_id: {type: string}
...
    MetaMatchPlayerLoadoutResponse:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/MetaMatchPlayerLoadout"}
        request_id: {type: string}
    MetaMatchCompleteRequest:
      type: object
      additionalProperties: false
      required: [result]
      properties:
        result: {type: object, additionalProperties: true}
...